Move consensus state object creation to ConsensusState.
Creation of the consensus state object for a particular block is better encapsulated inside the consensus state class instead of as a free-standing function in the common utils.
